Generate shared secret first instead of last
It seems that there is a race condition with Flight Hunter whereby, in
situations of automatic configuring of a cluster type with autoapply, at
some point between hunter restarting (and deciding to apply to the node)
and the stack restart that the profile process gets confused and dies
before properly establishing itself.

I can't see any reason why we'd want to do this generation last where
other setup scripts could do important things, instead we'll do this
before anything else to ensure that a shared secret exists before any
services can be started.